Fix token refresh freeze in long-running services
Long-running services would freeze an entire server when the token
needed to be refreshed. This was due to how the request interception was
implemented. The call to `futures::executor::block_on` when implementing
a `tonic::Interceptor` was the culprit, as the function runs in a tokyo
runtime, and the call to `block_on` ends upe blocking the runtime.

The solution is to implement a tower middleware that supports async calls
and ruturns a `Future` instead. This way, the runtime is not blocked and
the token refresh can be done in the background.

The solution was inspired by the `bigtable_rs` project, specifically the
https://github.com/liufuyang/bigtable_rs/blob/main/bigtable_rs/src/auth_service.rs.
service, but with changes to use the same types as the underlying service.